feat: support for --replicaof flag and extended INFO command
