+++ app/types.ts
@@ -13,3 +13,4 @@ export type StoreValue = {
 export type StreamEntry = [string, string[]];
 export type EntryData = Record<string, string>;
 
+export type ServerRole = 'master' | 'slave';
